Made hdRelation thread safe (except for one function).
Changed behavior: if a file exists then an error is returned.
Reason: Sometimes the lib is used on the same topology again, this leads to multiple 
writers to a single file => file is broken.
To fix this issue O_EXCL is used, then the user gets an error!

Fixed also some mem leaks if there is an error during file open.